# Conflicts:
#	target/m2e-wtp/web-resources/META-INF/maven/Bartenders/Bartenders/pom.properties

+++ src/bar/controller/MessageBoardController.java
@@ -8,9 +8,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import javax.mail.Session;
-
-import org.apache.tomcat.jni.User;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.ui.Model;
@@ -21,15 +18,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.SessionAttributes;
 
 import bar.model.MessageBoard;
-import bar.model.MessageBoardDAO;
+
 import bar.model.MessageBoardService;
 import bar.model.SubMessageBoard;
-import bar.model.UsersDAO;
+
 import bar.model.UsersService;
-import bar.model.Users;
 
 @Controller
-@SessionAttributes({ "LoginStatus", "account" ,"userName"})
+@SessionAttributes({ "LoginStatus", "account", "userName" })
 @EnableTransactionManagement
 public class MessageBoardController {
 
@@ -50,13 +46,12 @@ public class MessageBoardController {
 		Map<String, String> errors = new HashMap<String, String>();
 		m.addAttribute("errors", errors);
 
-		if(account==null) {
+		if (account == null) {
 			return "index";
 		}
-		
 
 		String rightblabla = blabla.replaceAll("\n", "<br>");
-		
+
 		String userName = usersService.select(account).getUserName();
 
 		m.addAttribute("userName", userName);
@@ -67,19 +62,16 @@ public class MessageBoardController {
 		m.addAttribute("deletePassword", deletePassword);
 
 		String time = getDateTime();
-		
-		
-		
 
 		MessageBoard messageBoard = new MessageBoard(1, title, account, time, rightblabla, picture, deletePassword,
-				userName,0);
+				userName, 0);
 		messageBoardService.createMessage(messageBoard);
 
 		List<MessageBoard> newest = messageBoardService.selectNewestMessage();
 
 		m.addAttribute("newest", newest);
-		
-		//for websocket
+
+		// for websocket
 		WebSocketTest.setModel(m);
 
 		return "MessageBoard";
@@ -98,59 +90,87 @@ public class MessageBoardController {
 	public String processActionShow(Model m) {
 		List<MessageBoard> newest = messageBoardService.selectNewestMessage();
 		m.addAttribute("newest", newest);
-		
-		//for websocket
+
+		// for websocket
 		WebSocketTest.setModel(m);
-		
+
 		return "MessageBoard";
 	}
-	
+
 	@RequestMapping(path = { "messageBoardShowList.controller" })
 	public String processActionShowList(Model m) {
 		List<MessageBoard> newest = messageBoardService.selectNewestMessage();
-			
-		for (int i = 0; i < newest.size(); i++) {
+
+		for (int i = 0; i <= newest.size() - 1; i++) {
 			int id = newest.get(i).getId();
 			List<SubMessageBoard> subnewest = messageBoardService.selectNewestSubMessage(id);
 			int subMessageAmount = subnewest.size();
 			newest.get(i).setSubMessageAmount(subMessageAmount);
-		}		
-		Collections.sort(newest,
-		        new Comparator<MessageBoard>() {
-		            public int compare(MessageBoard o1, MessageBoard o2) {
-		                return o2.getSubMessageAmount()-o1.getSubMessageAmount();
-		            }
-		        });
-			
-		String famousMessageRank1PitcureUrl = newest.get(0).getPicture();	
-		String famousMessageRank2PitcureUrl = newest.get(1).getPicture();
-		String famousMessageRank3PitcureUrl = newest.get(2).getPicture();
-		
-		
-		m.addAttribute("famousMessageRank1PitcureUrl", famousMessageRank1PitcureUrl);		
-		m.addAttribute("famousMessageRank2PitcureUrl", famousMessageRank2PitcureUrl);
-		m.addAttribute("famousMessageRank3PitcureUrl", famousMessageRank3PitcureUrl);
-		
-		m.addAttribute("newest", newest);
-		return "MessageBoardListMode";
+		}
+		Collections.sort(newest, new Comparator<MessageBoard>() {
+			public int compare(MessageBoard o1, MessageBoard o2) {
+				return o2.getSubMessageAmount() - o1.getSubMessageAmount();
+			}
+		});
+
+		if (newest.size() < 3) {
+			m.addAttribute("famousMessageRank1title", "文章數量不夠");
+			m.addAttribute("famousMessageRank2title", "文章數量不夠");
+			m.addAttribute("famousMessageRank3title", "文章數量不夠");
+			m.addAttribute("newest", newest);
+
+			return "MessageBoardListMode";
+		} else {
+			String famousMessageRank1PitcureUrl = newest.get(0).getPicture();
+			String famousMessageRank1title = newest.get(0).getTitle();
+
+			String famousMessageRank2PitcureUrl = newest.get(1).getPicture();
+			String famousMessageRank2title = newest.get(1).getTitle();
+
+			String famousMessageRank3PitcureUrl = newest.get(2).getPicture();
+			String famousMessageRank3title = newest.get(2).getTitle();
+
+			m.addAttribute("famousMessageRank1PitcureUrl", famousMessageRank1PitcureUrl);
+			m.addAttribute("famousMessageRank1title", famousMessageRank1title);
+
+			m.addAttribute("famousMessageRank2PitcureUrl", famousMessageRank2PitcureUrl);
+			m.addAttribute("famousMessageRank2title", famousMessageRank2title);
+
+			m.addAttribute("famousMessageRank3PitcureUrl", famousMessageRank3PitcureUrl);
+			m.addAttribute("famousMessageRank3title", famousMessageRank3title);
+
+			m.addAttribute("newest", newest);
+			return "MessageBoardListMode";
+		}
+
 	}
-	
-	
 
 	@RequestMapping(path = { "messageBoardDelete.controller" }, method = { RequestMethod.POST })
 	public String processActionDelete(@RequestParam(name = "id") int id,
-			@RequestParam(name = "deletePassword") String deletePassword, Model m,
-			@ModelAttribute(name = "account") String account) {
+			@RequestParam(name = "deletePassword") String deletePassword, Model m) {
 
-		boolean status = messageBoardService.delete(id, deletePassword);
+		MessageBoard theMessage = messageBoardService.selectTheMessageOnlyById(id);
+		
+		if(theMessage==null) {
 
-		if (status) {
 			List<MessageBoard> newest = messageBoardService.selectNewestMessage();
 
 			m.addAttribute("newest", newest);
 
 			return "MessageBoard";
 		}
+		
+		Integer amountSubMessage = theMessage.getSubMessageAmount();						
+		if (amountSubMessage==0) {			
+			messageBoardService.delete(id, deletePassword);
+			
+			List<MessageBoard> newest = messageBoardService.selectNewestMessage();
+
+			m.addAttribute("newest", newest);
+
+			return "MessageBoard";
+		}
+
 		List<MessageBoard> newest = messageBoardService.selectNewestMessage();
 
 		m.addAttribute("newest", newest);
